done update product shop and add new product option of product

+++ ede-be/ede-product-service/src/main/java/com/ede/edeproductservice/restcontroller/user/UpdateProductRestController.java
@@ -1,5 +1,7 @@
 package com.ede.edeproductservice.restcontroller.user;
 
+import java.util.UUID;
+
 import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -36,10 +38,10 @@ public class UpdateProductRestController {
 
 	@Autowired
 	Auth_Service auth_service;
-	
+
 	@Autowired
 	ShopService shopService;
-	
+
 	@Autowired
 	Product_brand_service brandService;
 
@@ -51,16 +53,20 @@ public class UpdateProductRestController {
 
 	@Autowired
 	Product_option_image_service product_option_image_service;
-	
+
 	@Autowired
 	Product_discount_service product_discount_service;
-	
+
 	@Autowired
 	Product_Tag_service product_Tag_service;
 
+	public String generateUUID() {
+		return UUID.randomUUID().toString();
+	}
+	
 	@SuppressWarnings("rawtypes")
 	@PutMapping("/update/product-shop/")
-	public ResponseEntity updateProduct(@RequestBody Product product, HttpServletRequest req){
+	public ResponseEntity updateProduct(@RequestBody Product product, HttpServletRequest req) {
 		User us = new User();
 		try {
 			us = auth_service.getUserLogin(req.getHeader("Authorization"));
@@ -72,14 +78,13 @@ public class UpdateProductRestController {
 		product.setShop(sh);
 		return ResponseEntity.status(HttpStatus.OK).body(service.save(product));
 	}
-	
+
 	@SuppressWarnings("rawtypes")
 	@PutMapping("/update/product-shop/discount")
 	public ResponseEntity updateProductDiscount(@RequestBody Product_discount discount) {
 		return ResponseEntity.status(HttpStatus.OK).body(product_discount_service.save(discount));
 	}
 
-	
 	@SuppressWarnings("rawtypes")
 	@PutMapping("/update/product-shop/options/")
 	public ResponseEntity updateProductOptions(@RequestBody Product_option product_option) {
@@ -90,12 +95,15 @@ public class UpdateProductRestController {
 	@SuppressWarnings("rawtypes")
 	@PutMapping("/update/product-shop/options/images")
 	public ResponseEntity updateProductOptionImage(@RequestBody Product_option_image product_option) {
-		product_option_image_service.deleteAllImage(product_option.getProductoption().getId());
-		String[] words = product_option.getImage().split(",");
-		for (int i = 0; i < words.length; i++) {
-			String[] fileCat = words[i].split("\\.");
-			product_option.setImage(product_option.getId() + "." + fileCat[1]);
-			product_option_image_service.save(product_option);
+		if (product_option != null) {
+			product_option_image_service.deleteAllImage(product_option.getProductoption().getId());
+			String[] words = product_option.getImage().split(",");
+			for (int i = 0; i < words.length; i++) {
+				String[] fileCat = words[i].split("\\.");
+				product_option.setId(generateUUID().toString());
+				product_option.setImage(product_option.getId() + "." + fileCat[1]);
+				product_option_image_service.save(product_option);
+			}
 		}
 		return ResponseHandler.generateResponse(HttpStatus.OK, true, "Cập nhật hình ảnh thành công", "", null);
 	}
@@ -103,10 +111,14 @@ public class UpdateProductRestController {
 	@SuppressWarnings("rawtypes")
 	@PutMapping("/update/product-shop/tag")
 	public ResponseEntity updateProductTag(@RequestBody Product_tag product_tag) {
-		String[] words = product_tag.getTag().split(",");
-		for (int i = 0; i < words.length; i++) {
-			product_tag.setTag(words[i]);
-			product_Tag_service.save(product_tag);
+		if (product_tag != null) {
+			product_Tag_service.deleteAllTagByIdProduct(product_tag.getProducttag().getId());
+			String[] words = product_tag.getTag().split(",");
+			for (int i = 0; i < words.length; i++) {
+				product_tag.setId(generateUUID().toString());
+				product_tag.setTag(words[i]);
+				product_Tag_service.save(product_tag);
+			}
 		}
 		return ResponseHandler.generateResponse(HttpStatus.OK, true, "Cập nhật tag thành công", "", null);
 	}
